Always use utf8 for the signatures
Technical note: The encoded header & payload should always be ascii.
So parsing ascii vs. parsing utf8 *shouldn't* make a difference semantically.
But, just to be sure it should definitely be the same (and not use ascii in one place and utf8 in another).
Plus, utf8 parsing is just much faster than ascii.